docs(site/ui): split xyflow Introduction into Nodes / Edges / Components
The Introduction grew into a one-page firehose. Split it along the
natural seams so each page tells one story:

  - /xyflow/introduction stays as the orientation page: overview,
    the createSignal+SVG simple example, when to reach for the
    package, install, the Hello/World Quick Start, and a Next
    Steps index to the new pages.

  - /xyflow/nodes is the deep dive on the node concept — the
    object shape, the default body, custom bodies via `nodeTypes`,
    and per-node `<Handle>` placement.

  - /xyflow/edges is the deep dive on edges — the object shape,
    default routing, the path-type variants `<SimpleEdge>` reads
    from `edge.type`, animated strokes / markers, and custom edges
    via `edgeTypes`.

  - /xyflow/components replaces the old `/components/xyflow`
    reference page. It carries the at-a-glance built-in component
    index plus the prop tables for `<Flow>` / `<Background>` /
    `<Controls>` / `<MiniMap>` / `<Handle>` / `<NodeWrapper>` /
    `<SimpleEdge>` (the prop tables grew to cover Handle /
    NodeWrapper / SimpleEdge too).

Sidebar nav, page prev/next (`xyflowOrder`), and the
`graph-editor` e2e spec now point at the new URLs. The old
`/components/xyflow` route and page are deleted — the URL 404s
intentionally so anything linking to the legacy path surfaces
in CI / monitoring.
